PYTHON-4530 - Synchronized test code should be in top-level test directory
