+++ .github/workflows/test-python.yml
@@ -206,5 +206,4 @@ jobs:
             which python
             pip install -e ".[test]"
             pytest -v
-            pytest -v test/synchronous/
             pytest -v test/asynchronous/
