+++ mypy_test.ini
@@ -7,7 +7,7 @@ exclude = (?x)(
     | ^test/conftest.py$
   )
 
-[mypy-pymongo.synchronous.*,gridfs.synchronous.*,test.synchronous.*]
+[mypy-pymongo.synchronous.*,gridfs.synchronous.*,test.*]
 warn_unused_ignores = false
 disable_error_code = unused-coroutine
