+++ test/conftest.py
@@ -4,6 +4,8 @@ from test import setup, teardown
 
 import pytest
 
+_IS_SYNC = True
+
 
 @pytest.fixture(scope="session", autouse=True)
 def test_setup_and_teardown():
