+++ tools/synchro.sh
@@ -1,5 +1,5 @@
 #!/bin/bash -eu
 
 python ./tools/synchro.py
-python -m ruff check pymongo/synchronous/ gridfs/synchronous/ test/synchronous --fix --silent
-python -m ruff format pymongo/synchronous/ gridfs/synchronous/ test/synchronous --silent
+python -m ruff check pymongo/synchronous/ gridfs/synchronous/ test/ --fix --silent
+python -m ruff format pymongo/synchronous/ gridfs/synchronous/ test/ --silent
